requirements.txt
anyio==3.7.1
boto3==1.28.73
botocore==1.31.73
certifi==2023.7.22
cffi==1.16.0
charset-normalizer==3.3.0
click==8.1.7
cryptography==41.0.4
ecdsa==0.18.0
fastapi==0.103.2
h11==0.14.0
idna==3.4
jmespath==1.0.1
psycopg2-binary==2.9.9
pyasn1==0.5.0
pycparser==2.21
pydantic==1.10.13
python-dateutil==2.8.2
python-dotenv==1.0.0
python-jose==3.3.0
requests==2.31.0
rsa==4.9
s3transfer==0.7.0
six==1.16.0
sniffio==1.3.0
SQLAlchemy==1.4.41
sqlalchemy2-stubs==0.0.2a35
sqlmodel==0.0.8
starlette==0.27.0
typing_extensions==4.8.0
urllib3==2.0.6
uvicorn==0.23.2
